Add simple signup link
The signup link is a mailto anchor tag, sending a message to the supplied email will automatically signup the sender for our mailing list.
To be replaced by a PHP form in the future. This is a minimum viable implementation as a result

Add styles for signup section by centering text
Add small change to text for signup

added margin-bottom to space out the signup row